Reorder APIs on the sidebar (#1122)
Since it is not in alphabetical order anyway, as far as I know "Polkadart" should be before "Polkadot-API"

Dart, comes before TypeScript, Python comes before TypeScript, and so on, but it seems the only "error" is that Dedot is listed first. 

If you don't mind us making this change @sinzii ? I really appreciate all your hard work.